Bug #1 FIXED
Added onto the solution document that explains the path to finding the solution to bug #1. Additionally, in the js file, the code has been added to. Added a 'if' statement for the firstElementChild and exited the function early using 'return'.

This fixed bug #1, so now no drop slots will hold more than one puzzle piece at a time