Bug #2 Fix
Added solution plan to to the "fixing_bugs_plan" rtf.

Fixed Bug #2 by adding a query selector "pieceContainer" and using it in the new function "resetPieces". The function creates a loop for each dropZone, to check for the condition 'if'.

The condition outlines that 'if' a child element is found in the dropZone, it should be appended to the variable pieceContainer.

Added the resetPieces callback into the function "changeBGImage" and fixed bug#2